backtesting_framework/pyproject.toml
[tool.poetry]
name = "backtesting-framework"
version = "0.1.0"
description = "Modular Financial Backtesting Framework"
authors = ["Financial Data Scientist <email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.0"
numpy = "^1.24.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
plotly = "^5.15.0"
quantstats = "^0.0.62"
scipy = "^1.11.0"
scikit-learn = "^1.3.0"
scikit-optimize = "^0.9.0"
pyyaml = "^6.0"
jinja2 = "^3.1.0"
click = "^8.1.0"
tqdm = "^4.65.0"
loguru = "^0.7.0"
yfinance = "^0.2.0"
empyrical = "^0.5.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
jupyter = "^1.0.0"

# Vectorbt removed due to Python 3.11+ compatibility issues
# Can be installed manually if needed: pip install vectorbt

[tool.poetry.scripts]
backtest = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true 